# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for program implementation MCDM methods TOPSIS and SMART",
        "title": "Decision Maker APi",
        "contact": {}
    },
    "host": "localhost:3030",
    "basePath": "/",
    "paths": {
...
        "/solution/final": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets final report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "final"
                ],
                "summary": "GetFinal",
                "operationId": "get-final",
                "parameters": [
                    {
                        "description": "threshold for sensitivity analysis",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controller.ThresholdInput"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "task identifier",
                        "name": "sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.FinalModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            }
        },
        "/solution/rating": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "gets ratings on current alternative in matrix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "GetRatings",
                "operationId": "get-ratings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "task identifier",
                        "name": "sid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "order of alternative",
                        "name": "ord",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.RatingsInput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "updates ratings in matrix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "UpdateMatrix",
                "operationId": "update-matrix",
                "parameters": [
                    {
                        "description": "ratings to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.RatingsInput"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "task identifier",
                        "name": "sid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "order of alternative",
                        "name": "ord",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "creates new matrix for evaluating",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "CreateMatrix",
                "operationId": "create-matrix",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "task identifier",
                        "name": "sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            }
        },
        "/solution/rating/complete": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "set status complete for current matrix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "CompleteStatus",
                "operationId": "set-complete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "task identifier",
                        "name": "sid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.RatingsInput": {
            "type": "object",
            "properties": {
                "ratings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eval.Rating"
                    }
                }
            }
        },
...
        "controller.ThresholdInput": {
            "type": "object",
            "properties": {
                "threshold": {
                    "type": "number"
                }
            }
        },
...
        "controller.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "entity.FinalModel": {
            "type": "object",
            "properties": {
                "fid": {
                    "type": "integer"
                },
                "last_change": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eval.Rating"
                    }
                },
                "sens_analysis": {
                    "$ref": "#/definitions/lib.SensitivityResult"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
...
        "eval.Rating": {
            "type": "object",
            "properties": {
                "eval": {}
            }
        },
        "lib.SensitivityResult": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/eval.Rating"
                        }
                    }
                },
                "threshold": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}